Move JWT queries out of GeneralQueries
